refactor(foundation): move SessionManager + drop allow(deprecated) from mockforge-vbr
Phase 6 / Sub-phase A7 follow-up: Move the full session.rs from
intelligent_behavior to foundation, which includes SessionManager,
SessionTracking, SessionTrackingMethod (previously only SessionState was moved).

Foundation now has:
- intelligent_behavior::session_state::{SessionState, InteractionRecord}
- intelligent_behavior::session::{SessionManager, SessionTracking, SessionTrackingMethod}

Added tokio + uuid workspace deps to foundation.

Removed #![allow(deprecated)] from mockforge-vbr — 359 tests pass + clippy
-D warnings clean. This is the first consumer crate to fully drop its
deprecation allow since Phase 6 began.

Also cleaned up unused imports in core (thiserror::Error, std::collections::HashMap
top-level, chrono::{DateTime, Utc}) that became unused after types moved to foundation.
Kept HashMap in config.rs tests module.
