+++ crates/mockforge-core/src/intelligent_behavior/config.rs
@@ -1,7 +1,6 @@
 //! Configuration for the Intelligent Mock Behavior system
 
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use std::time::Duration;
 
 use super::session::SessionTracking;
@@ -238,6 +237,7 @@ fn default_true() -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::collections::HashMap;
 
     #[test]
     fn test_config_defaults() {
