+++ crates/mockforge-foundation/src/intelligent_behavior/session.rs
@@ -1,198 +1,306 @@
-//! Session types: `InteractionRecord` and `SessionState`
-//!
-//! Extracted from `mockforge-core::intelligent_behavior::types` (Phase 6 / A7).
-//! Uses `crate::clock::now()` which honors any registered time-travel clock.
+//! Session management for tracking state across requests
 
-use crate::clock::now as clock_now;
-use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use std::sync::Arc;
+use tokio::sync::RwLock;
+use uuid::Uuid;
 
-/// A single interaction record (request + response pair)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct InteractionRecord {
-    /// Timestamp of the interaction
-    pub timestamp: DateTime<Utc>,
+use super::session_state::SessionState;
+use crate::Result;
 
-    /// HTTP method (GET, POST, PUT, DELETE, etc.)
-    pub method: String,
-
-    /// Request path (e.g., /api/users/123)
-    pub path: String,
-
-    /// Query parameters
-    #[serde(default)]
-    pub query_params: HashMap<String, String>,
-
-    /// Request headers
-    #[serde(default)]
-    pub headers: HashMap<String, String>,
-
-    /// Request body (if present)
-    pub request: Option<serde_json::Value>,
-
-    /// Response status code
-    pub status: u16,
-
-    /// Response body (if present)
-    pub response: Option<serde_json::Value>,
-
-    /// Vector embedding for semantic search (generated from interaction summary)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub embedding: Option<Vec<f32>>,
-
-    /// Metadata about this interaction
-    #[serde(default)]
-    pub metadata: HashMap<String, String>,
+/// Session tracking method
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[serde(rename_all = "lowercase")]
+#[derive(Default)]
+pub enum SessionTrackingMethod {
+    /// Track via cookie
+    #[default]
+    Cookie,
+    /// Track via HTTP header
+    Header,
+    /// Track via query parameter
+    QueryParam,
 }
 
-impl InteractionRecord {
-    /// Create a new interaction record (timestamped at the current wall-clock time).
-    pub fn new(
-        method: impl Into<String>,
-        path: impl Into<String>,
-        request: Option<serde_json::Value>,
-        status: u16,
-        response: Option<serde_json::Value>,
-    ) -> Self {
+/// Session tracking configuration
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct SessionTracking {
+    /// Tracking method
+    #[serde(default)]
+    pub method: SessionTrackingMethod,
+
+    /// Cookie name (if method is Cookie)
+    #[serde(default = "default_cookie_name")]
+    pub cookie_name: String,
+
+    /// Header name (if method is Header)
+    #[serde(default = "default_header_name")]
+    pub header_name: String,
+
+    /// Query parameter name (if method is QueryParam)
+    #[serde(default = "default_query_param")]
+    pub query_param: String,
+
+    /// Automatically create sessions if not present
+    #[serde(default = "default_true")]
+    pub auto_create: bool,
+}
+
+impl Default for SessionTracking {
+    fn default() -> Self {
         Self {
-            timestamp: Utc::now(),
-            method: method.into(),
-            path: path.into(),
-            query_params: HashMap::new(),
-            headers: HashMap::new(),
-            request,
-            status,
-            response,
-            embedding: None,
-            metadata: HashMap::new(),
+            method: SessionTrackingMethod::Cookie,
+            cookie_name: default_cookie_name(),
+            header_name: default_header_name(),
+            query_param: default_query_param(),
+            auto_create: true,
+        }
+    }
+}
+
+fn default_cookie_name() -> String {
+    "mockforge_session".to_string()
+}
+
+fn default_header_name() -> String {
+    "X-Session-ID".to_string()
+}
+
+fn default_query_param() -> String {
+    "session_id".to_string()
+}
+
+fn default_true() -> bool {
+    true
+}
+
+/// Session manager for tracking and managing sessions
+pub struct SessionManager {
+    /// Active sessions
+    sessions: Arc<RwLock<HashMap<String, SessionState>>>,
+
+    /// Session tracking configuration
+    config: SessionTracking,
+
+    /// Session timeout in seconds
+    timeout_seconds: u64,
+}
+
+impl SessionManager {
+    /// Create a new session manager
+    pub fn new(config: SessionTracking, timeout_seconds: u64) -> Self {
+        Self {
+            sessions: Arc::new(RwLock::new(HashMap::new())),
+            config,
+            timeout_seconds,
         }
     }
 
-    /// Add query parameters.
-    #[must_use]
-    pub fn with_query_params(mut self, params: HashMap<String, String>) -> Self {
-        self.query_params = params;
-        self
+    /// Generate a new session ID
+    pub fn generate_session_id() -> String {
+        Uuid::new_v4().to_string()
     }
 
-    /// Add headers.
-    #[must_use]
-    pub fn with_headers(mut self, headers: HashMap<String, String>) -> Self {
-        self.headers = headers;
-        self
+    /// Get or create a session
+    pub async fn get_or_create_session(&self, session_id: Option<String>) -> Result<String> {
+        let session_id = match session_id {
+            Some(id) => {
+                // Check if session exists
+                let sessions = self.sessions.read().await;
+                if sessions.contains_key(&id) {
+                    id
+                } else if self.config.auto_create {
+                    drop(sessions); // Release read lock
+                    let new_id = id.clone();
+                    self.create_session(new_id.clone()).await?;
+                    new_id
+                } else {
+                    return Err(crate::Error::internal(format!("Session '{}' not found", id)));
+                }
+            }
+            None => {
+                if self.config.auto_create {
+                    let new_id = Self::generate_session_id();
+                    self.create_session(new_id.clone()).await?;
+                    new_id
+                } else {
+                    return Err(crate::Error::internal(
+                        "No session ID provided and auto-create is disabled",
+                    ));
+                }
+            }
+        };
+
+        Ok(session_id)
     }
 
-    /// Set embedding.
-    #[must_use]
-    pub fn with_embedding(mut self, embedding: Vec<f32>) -> Self {
-        self.embedding = Some(embedding);
-        self
-    }
+    /// Create a new session
+    pub async fn create_session(&self, session_id: String) -> Result<String> {
+        let mut sessions = self.sessions.write().await;
 
-    /// Add metadata.
-    #[must_use]
-    pub fn with_metadata(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
-        self.metadata.insert(key.into(), value.into());
-        self
-    }
-
-    /// Generate a textual summary of this interaction for embedding.
-    pub fn summary(&self) -> String {
-        let request_body = self
-            .request
-            .as_ref()
-            .map(|r| serde_json::to_string(r).unwrap_or_default())
-            .unwrap_or_default();
-
-        let response_body = self
-            .response
-            .as_ref()
-            .map(|r| serde_json::to_string(r).unwrap_or_default())
-            .unwrap_or_default();
-
-        format!(
-            "{} {} | Request: {} | Status: {} | Response: {}",
-            self.method, self.path, request_body, self.status, response_body
-        )
-    }
-}
-
-/// Session state snapshot
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SessionState {
-    /// Session ID
-    pub session_id: String,
-
-    /// Current state data (e.g., logged-in user, cart items, etc.)
-    pub state: HashMap<String, serde_json::Value>,
-
-    /// Interaction history for this session
-    pub history: Vec<InteractionRecord>,
-
-    /// Session creation time
-    pub created_at: DateTime<Utc>,
-
-    /// Last activity time
-    pub last_activity: DateTime<Utc>,
-
-    /// Session metadata
-    #[serde(default)]
-    pub metadata: HashMap<String, String>,
-}
-
-impl SessionState {
-    /// Create a new session state.
-    ///
-    /// Uses `crate::clock::now()` which respects any registered virtual clock
-    /// (see `mockforge_foundation::clock::set_clock`).
-    pub fn new(session_id: impl Into<String>) -> Self {
-        let now = clock_now();
-        Self {
-            session_id: session_id.into(),
-            state: HashMap::new(),
-            history: Vec::new(),
-            created_at: now,
-            last_activity: now,
-            metadata: HashMap::new(),
+        if sessions.contains_key(&session_id) {
+            return Err(crate::Error::internal(format!("Session '{}' already exists", session_id)));
         }
+
+        let state = SessionState::new(session_id.clone());
+        sessions.insert(session_id.clone(), state);
+
+        Ok(session_id)
     }
 
-    /// Update last activity timestamp.
-    ///
-    /// Uses `crate::clock::now()` which respects any registered virtual clock.
-    pub fn touch(&mut self) {
-        self.last_activity = clock_now();
+    /// Get a session by ID
+    pub async fn get_session(&self, session_id: &str) -> Option<SessionState> {
+        let sessions = self.sessions.read().await;
+        sessions.get(session_id).cloned()
     }
 
-    /// Add an interaction to history.
-    pub fn record_interaction(&mut self, interaction: InteractionRecord) {
-        self.history.push(interaction);
-        self.touch();
+    /// Update a session
+    pub async fn update_session(&self, session_id: &str, state: SessionState) -> Result<()> {
+        let mut sessions = self.sessions.write().await;
+
+        if !sessions.contains_key(session_id) {
+            return Err(crate::Error::internal(format!("Session '{}' not found", session_id)));
+        }
+
+        sessions.insert(session_id.to_string(), state);
+        Ok(())
     }
 
-    /// Get a value from state.
-    pub fn get(&self, key: &str) -> Option<&serde_json::Value> {
-        self.state.get(key)
+    /// Delete a session
+    pub async fn delete_session(&self, session_id: &str) -> Result<()> {
+        let mut sessions = self.sessions.write().await;
+        sessions.remove(session_id);
+        Ok(())
     }
 
-    /// Set a value in state.
-    pub fn set(&mut self, key: impl Into<String>, value: serde_json::Value) {
-        self.state.insert(key.into(), value);
-        self.touch();
+    /// List all active session IDs
+    pub async fn list_sessions(&self) -> Vec<String> {
+        let sessions = self.sessions.read().await;
+        sessions.keys().cloned().collect()
     }
 
-    /// Remove a value from state.
-    pub fn remove(&mut self, key: &str) -> Option<serde_json::Value> {
-        let result = self.state.remove(key);
-        self.touch();
-        result
+    /// Clean up expired sessions
+    pub async fn cleanup_expired_sessions(&self) -> usize {
+        let timeout = chrono::Duration::seconds(self.timeout_seconds as i64);
+        let mut sessions = self.sessions.write().await;
+
+        let expired: Vec<String> = sessions
+            .iter()
+            .filter(|(_, state)| state.is_inactive(timeout))
+            .map(|(id, _)| id.clone())
+            .collect();
+
+        let count = expired.len();
+        for id in expired {
+            sessions.remove(&id);
+        }
+
+        count
     }
 
-    /// Check if session has been inactive for a duration.
-    ///
-    /// Uses `crate::clock::now()` which respects any registered virtual clock.
-    pub fn is_inactive(&self, duration: chrono::Duration) -> bool {
-        clock_now().signed_duration_since(self.last_activity) > duration
+    /// Get the number of active sessions
+    pub async fn session_count(&self) -> usize {
+        let sessions = self.sessions.read().await;
+        sessions.len()
+    }
+
+    /// Clear all sessions
+    pub async fn clear_all(&self) {
+        let mut sessions = self.sessions.write().await;
+        sessions.clear();
+    }
+
+    /// Get session tracking configuration
+    pub fn config(&self) -> &SessionTracking {
+        &self.config
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_session_manager_create_session() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 3600);
+
+        let session_id = manager.create_session("test_session".to_string()).await.unwrap();
+        assert_eq!(session_id, "test_session");
+
+        let state = manager.get_session(&session_id).await;
+        assert!(state.is_some());
+    }
+
+    #[tokio::test]
+    async fn test_session_manager_get_or_create() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 3600);
+
+        // Create with auto-create
+        let session_id = manager.get_or_create_session(None).await.unwrap();
+        assert!(!session_id.is_empty());
+
+        // Get existing
+        let same_id = manager.get_or_create_session(Some(session_id.clone())).await.unwrap();
+        assert_eq!(session_id, same_id);
+    }
+
+    #[tokio::test]
+    async fn test_session_manager_delete_session() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 3600);
+
+        let session_id = manager.create_session("test_delete".to_string()).await.unwrap();
+        assert!(manager.get_session(&session_id).await.is_some());
+
+        manager.delete_session(&session_id).await.unwrap();
+        assert!(manager.get_session(&session_id).await.is_none());
+    }
+
+    #[tokio::test]
+    async fn test_session_manager_list_sessions() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 3600);
+
+        manager.create_session("session1".to_string()).await.unwrap();
+        manager.create_session("session2".to_string()).await.unwrap();
+
+        let sessions = manager.list_sessions().await;
+        assert_eq!(sessions.len(), 2);
+        assert!(sessions.contains(&"session1".to_string()));
+        assert!(sessions.contains(&"session2".to_string()));
+    }
+
+    #[tokio::test]
+    async fn test_session_manager_clear_all() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 3600);
+
+        manager.create_session("session1".to_string()).await.unwrap();
+        manager.create_session("session2".to_string()).await.unwrap();
+
+        assert_eq!(manager.session_count().await, 2);
+
+        manager.clear_all().await;
+        assert_eq!(manager.session_count().await, 0);
+    }
+
+    #[tokio::test]
+    async fn test_session_cleanup_expired() {
+        let config = SessionTracking::default();
+        let manager = SessionManager::new(config, 1); // 1 second timeout
+
+        let session_id = manager.create_session("test_expire".to_string()).await.unwrap();
+
+        // Wait for session to expire
+        tokio::time::sleep(tokio::time::Duration::from_secs(2)).await;
+
+        let cleaned = manager.cleanup_expired_sessions().await;
+        assert_eq!(cleaned, 1);
+        assert!(manager.get_session(&session_id).await.is_none());
     }
 }
